refactor(devs-infra): bring back a part of old structure
