+++ .eslintignore
@@ -1,7 +1,7 @@
-src/build/
-src/setup-jest.ts
+build/
 node_modules/
 e2e/
+src/config/setup-jest.ts
 coverage/
 docs/
 .eslintrc.js
