+++ .npmignore
@@ -1,0 +1,43 @@
+# Tests
+e2e
+
+# sources are inlined
+src
+
+# Developement scripts
+scripts
+
+# Coverage directory used by tools like istanbul
+coverage
+
+# Dependency directories
+node_modules
+
+# Editor config
+.vscode
+.idea
+.editorconfig
+
+# CI
+.circleci
+
+# Eslint
+.eslintignore
+.eslintrc.js
+commitlint.config.js
+.prettierrc
+
+# Internal jest config
+jest.config.js
+
+# Tsconfig
+tsconfig.json
+tsconfig.build.json
+tsconfig.spec.json
+
+# Others
+*.tgz
+.npmignore
+.npmrc
+.github
+.gitattributes
