+++ e2e/test-app-v10-zone-v11/package.json
@@ -20,7 +20,7 @@
     "@angular/router": "~10.0.8",
     "rxjs": "~6.6.2",
     "tslib": "^2.0.1",
-    "zone.js": "~0.11.1"
+    "zone.js": "~0.10.3"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.1000.5",
@@ -30,7 +30,7 @@
     "@types/node": "^12.12.54",
     "codelyzer": "^6.0.0",
     "jest": "^26.3.0",
-    "jest-preset-angular": "file:../../src",
+    "jest-preset-angular": "file:../../build",
     "ts-node": "~8.10.2",
     "tslint": "~6.1.3",
     "typescript": "~3.9.7"
