+++ e2e/test-app-v10/package.json
@@ -30,7 +30,7 @@
     "@types/node": "^12.12.54",
     "codelyzer": "^6.0.0",
     "jest": "^26.3.0",
-    "jest-preset-angular": "file:../../src",
+    "jest-preset-angular": "file:../../build",
     "ts-node": "~8.10.2",
     "tslint": "~6.1.3",
     "typescript": "~3.9.7"
