+++ e2e/test-app-v9/package.json
@@ -34,7 +34,7 @@
     "@types/jest": "^26.0.9",
     "@types/node": "^14.0.1",
     "jest": "^26.3.0",
-    "jest-preset-angular": "file:../../src",
+    "jest-preset-angular": "file:../../build",
     "ts-node": "^8.10.2",
     "typescript": "~3.8.3"
   }
