--- src/.npmignore
@@ -1,6 +1,0 @@
-# Tests
-__tests__
-
-# sources are inlined
-*.ts
-*.tgz
