+++ src/index.ts
@@ -1,0 +1,1 @@
+export = require('./config/setup-jest');
