+++ tsconfig.json
@@ -10,6 +10,7 @@
     "importsNotUsedAsValues": "error",
     "skipLibCheck": true,
     "importHelpers": true,
+    "esModuleInterop": true,
     "target": "es2015",
     "module": "CommonJS",
     "lib": ["esnext", "dom"],
